requirements.txt
absl-py==2.0.0
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.1
beautifulsoup4==4.11.2
blinker==1.6.2
boto3==1.37.12
botocore==1.37.12
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cryptography==40.0.2
cycler==0.12.1
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
dj-database-url==2.3.0
Django==5.1.2
django-allauth==65.4.1
django-on-heroku==1.1.2
django-storages==1.14.5
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.44.3
frozendict==2.3.8
fsspec==2023.10.0
future==1.0.0
gast==0.5.4
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio==1.59.3
gunicorn==23.0.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.17.3
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
lxml==4.9.2
Markdown==3.5.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.0.7
multitasking==0.0.11
mutagen==1.47.0
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.7
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.18.0
pooch==1.8.0
proto-plus==1.26.1
protobuf==3.20.3
psycopg==3.2.6
psycopg2-binary==2.9.10
pyarrow==12.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.3.0
Pygments==2.15.1
PyJWT==2.10.1
Pympler==1.0.1
pyparsing==3.1.1
pypng==0.20220715.0
pyrsistent==0.19.3
PySimpleGUI==4.60.4
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.3.5
rsa==4.9
s3transfer==0.11.4
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.7
sqlparse==0.4.3
streamlit==1.22.0
stripe==11.6.0
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
textblob==0.17.1
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==5.0.1
urllib3==1.26.14
validators==0.20.0
virtualenv==20.20.0
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==2.2.3
whitenoise==6.9.0
wrapt==1.14.1
yfinance==0.2.18
zipp==3.15.0
